Move TimestampDigest into interfaces
